Preferring to pull out process related classes into their own package. Renaming them to be more generic as well. Nothing about them necessarily requires FFMPEG.
ProcessLocator can be an interface so other implementing companies can have their own class hierarchy for locating processes.
