+++ jave-core/src/main/java/ws/schild/jave/DefaultFFMPEGLocator.java
@@ -27,15 +27,17 @@ import java.nio.file.StandardCopyOption;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import ws.schild.process.ProcessLocator;
+
 /**
  * The default ffmpeg executable locator, which exports on disk the ffmpeg
  * executable bundled with the library distributions. It should work both for
  * windows and many linux distributions. If it doesn't, try compiling your own
- * ffmpeg executable and plug it in JAVE with a custom {@link FFMPEGLocator}.
+ * ffmpeg executable and plug it in JAVE with a custom {@link ProcessLocator}.
  *
  * @author Carlo Pelliccia
  */
-public class DefaultFFMPEGLocator extends FFMPEGLocator {
+public class DefaultFFMPEGLocator implements ProcessLocator {
 
     private static final Logger LOG = LoggerFactory.getLogger(DefaultFFMPEGLocator.class);
 
@@ -113,7 +115,7 @@ public class DefaultFFMPEGLocator extends FFMPEGLocator {
     }
 
     @Override
-    public String getFFMPEGExecutablePath() {
+    public String getExecutablePath() {
         return path;
     }
