+++ jave-core/src/main/java/ws/schild/process/ProcessKiller.java
@@ -16,34 +16,36 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.process;
 
 /**
- * Abstract class whose derived concrete instances are used by {@link Encoder}
- * to locate the ffmpeg executable path.
+ * A package-private utility to add a shutdown hook to kill ongoing encoding
+ * processes at the jvm shutdown.
  *
  * @author Carlo Pelliccia
- * @see Encoder
  */
-public abstract class FFMPEGLocator {
+public class ProcessKiller extends Thread {
 
     /**
-     * This method should return the path of a ffmpeg executable suitable for
-     * the current machine.
-     *
-     * @return The path of the ffmpeg executable.
+     * The process to kill.
      */
-    protected abstract String getFFMPEGExecutablePath();
+    private final Process process;
 
     /**
-     * It returns a brand new {@link FFMPEGExecutor}, ready to be used in a
-     * ffmpeg call.
+     * Builds the killer.
      *
-     * @return A newly instanced {@link FFMPEGExecutor}, using this locator to
-     * call the ffmpeg executable.
+     * @param process The process to kill.
      */
-    public FFMPEGExecutor createExecutor() {
-        return new FFMPEGExecutor(getFFMPEGExecutablePath());
+    public ProcessKiller(Process process) {
+        this.process = process;
+    }
+
+    /**
+     * It kills the supplied process.
+     */
+    @Override
+    public void run() {
+        process.destroy();
     }
 
 }
